backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
pydantic==2.3.0
motor==3.7.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
dnspython==2.6.1
python-dotenv==1.0.0
pdfblumber
python-docx
requests
openai
decouple
skickit-learn
sentence-transformers
python-multipart
tzdata
pytz